pyproject.toml
[project]
name = "slapdash"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi-sessions>=0.3.2",
    "fastapi[standard]>=0.117.1",
    "jinja2-fragments>=1.10.0",
    "mypy>=1.18.2",
    "pyright>=1.1.406",
    "requests>=2.32.5",
    "sqlmodel>=0.0.25",
    "uvicorn>=0.37.0",
    "websockets>=15.0.1",
]
indentation = 2

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.ruff]
indent-width = 2

[tool.ruff.format]
indent-style = "space"

[dependency-groups]
dev = [
    "ipython>=9.6.0",
]
